🔧 build : 프로필 수정 완료

See merge request s05-webmobile2-sub3/S05P13D104!165

+++ backend/webcuration/src/main/java/com/web/webcuration/service/UserServiceImpl.java
@@ -3,6 +3,7 @@ package com.web.webcuration.service;
 import java.io.IOException;
 import java.util.Optional;
 
+import com.web.webcuration.Entity.Provide;
 import com.web.webcuration.Entity.User;
 import com.web.webcuration.dto.request.NickNameRequest;
 import com.web.webcuration.dto.request.ProfileRequest;
@@ -67,7 +68,13 @@ public class UserServiceImpl implements UserService {
                     // 사진이 변경 됐으면
                     if (!changeUser.getImage().equals("/profileImg/user_image.png")) {
                         // 기본 프로필이 아니면
-                        FileUtils.deleteProfile(changeUser.getImage());
+                        if (changeUser.getProvide().equals(Provide.LOCAL)) {
+                            // 로컬이면
+                            FileUtils.deleteProfile(changeUser.getImage());
+                        } else {
+                            // SNS면
+                            changeUser.setProvide(Provide.LOCAL);
+                        }
                     }
                     if (profileRequest.getImage() == null) {
                         // 기본 프로필 설정을 했으면
